requirements.txt
web3
pandas
pyyaml
deepdiff